requirements.txt
 python==3.7.10  # Look like this is the only one that does not conflict with prokka 1.14.6
 ipykernel
 matplotlib
 numpy
 pandas
 scipy
 tqdm
 xlrd
 parasail-python
 biopython
 pysam
 spades==3.15.2
 prokka==1.14.6
 mlst==2.19.0
 abricate==1.0.1
 shovill==1.1.0
 roary==3.13.0
 iqtree==2.0.3
 sra-tools==2.11.0
 ete3==3.1.2
 kraken2==2.1.1
 fastqc==0.11.9
 multiqc==1.10
 quast==5.0.2
 ncbi-amrfinderplus==3.10.40
 isescan==1.7.2.3
 flye=2.9
 cd-hit==4.8.1
 blast==2.13.0
 diamond==2.1.3
 mcl==14.137
 mafft==7.515
 psutil
